+++ vjer/__main__.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+"""
+__main__ module for Vjer.
+
+Enables use as module: $ python -m vjer --version
+"""
+from . import vjer
+
+if __name__ == '__main__':
+    vjer.main()
+
+# cSpell:ignore vjer