+++ vjer/rollback.py
@@ -12,7 +12,7 @@ class RollbackStep(VjerStep):
 
     def rollback_helm(self) -> None:
         """Rollback method for Helm charts."""
-        chart_name = self.step_info.chart_name if self.step_info.chart_name else self.project.product.lower()
+        chart_name = self.step_info.chart_name if self.step_info.chart_name else self.project.name.lower()
         helm('rollback',
              self.step_info.release_name.lower() if self.step_info.release_name else chart_name,
              wait=True, **self.helm_args)